+++ src/client/java/com/provismet/cobblemon/gimmick/datagen/EnchantmentGenerator.java
@@ -1,6 +1,8 @@
 package com.provismet.cobblemon.gimmick.datagen;
 
+import com.provismet.cobblemon.gimmick.api.data.datapack.condition.EnchantmentToggleCondition;
 import com.provismet.cobblemon.gimmick.registry.GTGEnchantments;
+import com.provismet.cobblemon.gimmick.registry.GTGResourceConditions;
 import com.provismet.lilylib.datagen.provider.LilyEnchantmentProvider;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.minecraft.registry.RegistryWrapper;
@@ -14,9 +16,9 @@ public class EnchantmentGenerator extends LilyEnchantmentProvider {
 
     @Override
     protected void generate (RegistryWrapper.WrapperLookup wrapperLookup, EnchantmentBuilder enchantmentBuilder) {
-        enchantmentBuilder.add(GTGEnchantments.KEY_STONE);
-        enchantmentBuilder.add(GTGEnchantments.Z_RING);
-        enchantmentBuilder.add(GTGEnchantments.DYNAMAX_BAND);
-        enchantmentBuilder.add(GTGEnchantments.TERA_ORB);
+        enchantmentBuilder.add(GTGEnchantments.KEY_STONE, new EnchantmentToggleCondition());
+        enchantmentBuilder.add(GTGEnchantments.Z_RING, new EnchantmentToggleCondition());
+        enchantmentBuilder.add(GTGEnchantments.DYNAMAX_BAND, new EnchantmentToggleCondition());
+        enchantmentBuilder.add(GTGEnchantments.TERA_ORB, new EnchantmentToggleCondition());
     }
 }