+++ changelog.md
@@ -1,7 +1,9 @@
 Feature update to add more functionality surrounding datapacks.
 
 ## Additions
-- Added the ability to datapack conditions, typecharts, scripts and held items in data/gimme-that-gimmick/showdown
+- Added the ability to datapack conditions, typecharts, scripts and held items in data/gimme-that-gimmick/showdown.
+- Added Oricorio nectars.
+- Enchantments can now be disabled in the config.
 
 ## Bugfixes
 - Max Mushroom blocks now correctly output their age and max age for compatibility purposes.